refactor dashboard components into dashboard directory
