+++ app/client/dashboard/components/Dashboard.coffee
