+++ app/client/dashboard/components/MotorList/Motor.js
