+++ app/client/dashboard/components/MotorList/MotorTester.js
